+++ R/expand_model_out_val_grid.R
@@ -10,11 +10,22 @@
 #' contains only a single round.
 #' @param required_vals_only Logical. Whether to return only combinations of
 #' Task ID and related output type ID required values.
+#' @param all_character Logical. Whether to return all character column.
+#' @param bind_model_tasks Logical. Whether to bind expanded grids of
+#' values from multiple modeling tasks into a single tibble/arrow table or
+#' return a list.
 #'
-#' @return a tibble containing all possible task ID and related output type ID
-#' value combinations.
+#' @return If `bind_model_tasks = TRUE` (default) a tibble or arrow table
+#' containing all possible task ID and related output type ID
+#' value combinations. If `bind_model_tasks = FALSE`, a list containing a
+#' tibble or arrow table for each round modeling task.
+#'
+#' Columns are coerced to data types according to the hub schema,
+#' unless `all_character = TRUE`. If `all_character = TRUE`, all columns are returned as
+#' character which can be faster when large expanded grids are expected.
 #' If `required_vals_only = TRUE`, values are limited to the combinations of required
 #' values only.
+#' @inheritParams coerce_to_hub_schema
 #' @details
 #' When a round is set to `round_id_from_variable: true`,
 #' the value of the task ID from which round IDs are derived (i.e. the task ID
@@ -42,9 +53,23 @@
 #' expand_model_out_val_grid(config_tasks, round_id = "2022-10-01")
 #' # Later round_id maps to round config that includes additional task ID 'age_group'.
 #' expand_model_out_val_grid(config_tasks, round_id = "2022-10-29")
+#' # Coerce all columns to character
+#' expand_model_out_val_grid(config_tasks,
+#'   round_id = "2022-10-29",
+#'   all_character = TRUE
+#' )
+#' # Return arrow table
+#' expand_model_out_val_grid(config_tasks,
+#'   round_id = "2022-10-29",
+#'   all_character = TRUE,
+#'   as_arrow_table = TRUE
+#' )
 expand_model_out_val_grid <- function(config_tasks,
                                       round_id,
-                                      required_vals_only = FALSE) {
+                                      required_vals_only = FALSE,
+                                      all_character = FALSE,
+                                      as_arrow_table = FALSE,
+                                      bind_model_tasks = TRUE) {
   round_idx <- get_round_idx(config_tasks, round_id)
 
   round_config <- purrr::pluck(
@@ -77,20 +102,23 @@ expand_model_out_val_grid <- function(config_tasks,
     process_grid_inputs(required_vals_only = required_vals_only) %>%
     purrr::map(~ purrr::compact(.x))
 
-  output_type_grid_l <- purrr::map2(
+  # Expand output grid individually for each output type and coerce to hub schema
+  # data types.
+
+  purrr::map2(
     task_id_l, output_type_l,
     ~ expand_output_type_grid(
       task_id_values = .x,
       output_type_values = .y
     )
-  )
-
-  do.call(rbind, output_type_grid_l) %>%
-    tibble::as_tibble() %>%
-    coerce_to_hub_schema(config_tasks)
+  ) %>%
+    process_mt_grid_outputs(
+      config_tasks,
+      all_character = all_character,
+      as_arrow_table = as_arrow_table,
+      bind_model_tasks = bind_model_tasks
+    )
 }
-
-
 
 process_grid_inputs <- function(x, required_vals_only = FALSE) {
   if (required_vals_only) {
@@ -134,3 +162,69 @@ fix_round_id <- function(x, round_id, round_config, round_ids) {
   }
 }
 
+
+process_mt_grid_outputs <- function(x, config_tasks, all_character,
+                                    as_arrow_table = TRUE,
+                                    bind_model_tasks = TRUE) {
+  if (bind_model_tasks) {
+    # To bind multiple modeling task grids together, we need to ensure they contain
+    # the same columns. Any missing columns are padded with NAs.
+    all_cols <- purrr::map(x, ~ names(.x)) %>%
+      unlist() %>%
+      unique()
+
+    schema_cols <- names(
+      create_hub_schema(
+        config_tasks,
+        partitions = NULL
+      )
+    )
+    all_cols <- schema_cols[schema_cols %in% all_cols]
+    x <- purrr::map(x, ~ pad_missing_cols(.x, all_cols))
+  }
+
+  if (all_character) {
+    x <- purrr::map(
+      x, ~ coerce_to_character(
+        .x,
+        as_arrow_table = as_arrow_table
+      )
+    )
+  } else {
+    x <- purrr::map(
+      x,
+      ~ coerce_to_hub_schema(
+        .x,
+        config_tasks,
+        as_arrow_table = as_arrow_table
+      )
+    )
+  }
+  if (bind_model_tasks) {
+    return(do.call(rbind, x))
+  } else {
+    return(x)
+  }
+}
+
+
+pad_missing_cols <- function(x, all_cols) {
+  if (inherits(x, "data.frame")) {
+    x[, all_cols[!all_cols %in% names(x)]] <- NA
+    return(x[, all_cols])
+  }
+  if (inherits(x, "ArrowTabular")) {
+    missing_colnames <- setdiff(all_cols, names(x))
+    if (length(missing_colnames) == 0L) {
+      return(x)
+    }
+
+    missing_cols <- as.list(rep(NA, length(missing_colnames))) %>%
+      stats::setNames(missing_colnames) %>%
+      as.data.frame() %>%
+      arrow::arrow_table()
+
+    return(cbind(x, missing_cols)[, all_cols])
+  }
+  x
+}